backend/requirements.txt
Flask
psycopg2-binary
gunicorn
pytest